1995 to 2015 Data Cleaned
Need at least 30 observations per estimate, and there is not enough data for most of the estimates interested in making. Added 8 more years of clean data, and still not enough (30) cases to make reliable estimates. Planning on adding NHAMCS data as well to redress this issue. Also, 2016 data was recently released, need to add.