Fix: 2002 Drug Codes as Name Strings
The drug codes for 2002 were not included in the original 2002 do file from NCHS. The drug name codes from the 2003 do file were copied into the 2002 do file, and the .dta file for 2002 was reproduced. Then the updated 2002 dta file was ran through the 0.0 and 0.1 files in the pemphigus analysis to reproduce the 1995 to 2015 file with appropriate drug names. This was used to list appropriate drug names in the atopic dermatitis analysis.

Next: make a separate github for the data processing/wrangling of the NAMCS datasets. Then incorporate pemphigus, atopic dermatitis, and future analyses in subfolders on GitHub.